refactor(tonk-schema): host wire Query and Conclusion types
Move the on-the-wire `Query` and `Conclusion` from
`tonk-worker::reactor` into `tonk-schema::{query, conclusion}` and
add `Deserialize` to `Conclusion` so browser clients can parse
them. `tonk-worker::reactor` re-exports both for back-compat.
